requirements.txt
Flask
Pillow
Werkzeug
mysql-connector-python
python-dotenv
numpy
torch
torchvision
reportlab
scikit-learn

